requirements.txt
flask
torch
torchvision
transformers
ultralytics
pillow
